+++ app/vaccinationAgentMenu/farmsAndAnimals/animalDetails/[animalId].tsx
@@ -101,7 +101,7 @@ export default function AnimalDetails() {
                     onPress={() => router.back()}
                     activeOpacity={0.7}
                 >
-                    <Icon name="arrow-left" size={24} color={COLORS.forestGreen} />
+                    <Icon name="arrow-left" size={24} color={COLORS.white} />
                 </TouchableOpacity>
                 <Text style={styles.title}>Características</Text>
             </View>
@@ -152,16 +152,21 @@ const styles = StyleSheet.create({
         justifyContent: "center",
         marginBottom: 24,
         position: "relative",
+        backgroundColor: COLORS.forestGreen,
+        paddingHorizontal: 15,
+        paddingVertical: 15,
+        borderRadius: 8,
     },
     backButton: {
         position: "absolute",
         left: 0,
         padding: 8,
+        marginLeft: 10
     },
     title: {
         fontSize: 28,
         fontWeight: "700",
-        color: COLORS.forestGreen,
+        color: COLORS.white,
         letterSpacing: 0.5,
     },
     animalCard: {
