+++ app/vaccinationAgentMenu/farmsAndAnimals/farmDetails/[farmId].tsx
@@ -315,7 +315,7 @@ export default function FarmDetails() {
                     onPress={() => router.replace("/vaccinationAgentMenu/farmsAndAnimals/farmsView")}
                     activeOpacity={0.7}
                 >
-                    <Icon name="arrow-left" size={24} color={COLORS.forestGreen} />
+                    <Icon name="arrow-left" size={24} color={COLORS.white} />
                 </TouchableOpacity>
                 <Text style={styles.title}>Detalles</Text>
             </View>
@@ -430,16 +430,21 @@ const styles = StyleSheet.create({
         justifyContent: "center",
         marginBottom: 24,
         position: "relative",
+        backgroundColor: COLORS.forestGreen,
+        paddingHorizontal: 15,
+        paddingVertical: 15,
+        borderRadius: 8,
     },
     backButton: {
         position: "absolute",
         left: 0,
         padding: 8,
+        marginLeft: 10
     },
     title: {
         fontSize: 28,
         fontWeight: "700",
-        color: COLORS.forestGreen,
+        color: COLORS.white,
         letterSpacing: 0.5,
     },
     farmCard: {
